Add EncryptedAdapter to encrypt cookies
This takes advantage of the MessageEncryptor class to encrypt and
decrypt a cookie's value. The key remains unencrypted (perhaps
obviously) but the value is encrypted with the app's secret_key_base.

The key must be 32 digits long (e.x. 16 hex pairs) or longer to be
valid for the encryptor. Because of this, I changed the secert_key_base
in the spec_helper.

I also encoded the result of the message encryptor. Currently it
encrypts it and returns raw bytes. I thought having a string
interpretation would be nicer to look at and use. I looked into the
hexstring method but there was no built-in way for me to get back out
of the hexstring that I could find. Base64 worked great.