Cookie and Session are fetched lazily
When the cookies are accessed from an action, they are automatically
read and parsed from the request. They are memoized so as to only slow
down the request when necessary.

When the session is accessed from an action, is uses data from the
cookies to retrieves its contents. This will call the cookies as
describe above. The session will then also be memoized.