Refactor FlashStore so it erases it self each request
For some reason I thought I could have a single store, but I couldn't.
Now the flash has two stores, now and next. Now can only be set from a
value stored on a session. Next is a more traditional mutable store.
It's what will be changed when actions call flash.set(key, value).

Eventually, the flash will be saved to JSON and written out to the
session, which will in turn be written out as a cookie to the browser.